backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.10.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
